requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
causal-learn==0.1.3.6
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.1.0
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
et-xmlfile1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
fonttools==4.39.4
frozenlist==1.4.0
graphviz==0.20.1
gurobipy==10.0.2
idna==3.4
importlib-resources==5.12.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
joblib==1.2.0
JPype1==1.4.1
jupyter-client==7.3.4
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy_loader==0.2
littleutils==0.2.2
llvmlite==0.40.1rc1
markdown-it-py==3.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.0
numba==0.57.0
numpy==1.24.2
openai==0.28.0
openpyxl==3.1.2
outdated==0.2.2
packaging==23.1
pandas==1.5.3
pandas-flavor==0.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pingouin==0.5.3
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pydot==1.4.2
Pygments==2.16.1
pygobnilp==1.0
pygraphviz==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==19.0.2
requests==2.31.0
retry==0.9.2
rich13.5.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
tabulate==0.9.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
wheel==0.38.4
xarray==2023.5.0
yarl==1.9.2
zipp==3.15.0
